+++ components/templates/src/lib.rs
@@ -11,7 +11,7 @@ use tera::{Context, Tera};
 use crate::filters::{MarkdownFilter, NumFormatFilter};
 use crate::functions::{
     GetHash, GetImageMetadata, GetPage, GetSection, GetTaxonomy, GetTaxonomyTerm, GetTaxonomyUrl,
-    GetUrl, LoadData, ResizeImage, Trans,
+    GetUrl, LoadData, ResizeImage, TextDirection, Trans,
 };
 use config::Config;
 use errors::{Context as ErrorContext, Result, bail};
@@ -70,6 +70,7 @@ pub static ZOLA_TERA: LazyLock<Tera> = LazyLock::new(|| {
     tera.register_function("get_taxonomy_term", GetTaxonomyTerm::default());
     tera.register_function("get_taxonomy_url", GetTaxonomyUrl::default());
     tera.register_function("trans", Trans::default());
+    tera.register_function("text_direction", TextDirection::default());
     tera.register_function("resize_image", ResizeImage::default());
     tera.register_function("get_image_metadata", GetImageMetadata::default());
     tera.register_function("load_data", LoadData::default());